Leave root legacy dir itself as empty directory after delete (#5890)

+++ mediorum/cmd/reaper/reaper.go
@@ -9,14 +9,13 @@ import (
 	"time"
 )
 
-func deleteFilesAndEmptyDirs(path string, logFile *os.File) error {
-	log.SetOutput(logFile)
+var rootPath string
 
-	if _, exists := os.LookupEnv("NO_LEGACY_REAPER"); exists || strings.HasSuffix(os.Getenv("creatorNodeEndpoint"), ".audius.co") {
-		log.Printf("Skipping deletion due to NO_LEGACY_REAPER env var present.")
-		log.Printf("Sleeping.")
-		time.Sleep(10000 * time.Hour)
-		return nil
+func deleteFilesAndEmptyDirs(path string, logFile *os.File) error {
+
+	if rootPath == "" {
+		// set here so works with `go test`
+		rootPath = path
 	}
 
 	entries, err := os.ReadDir(path)
@@ -39,16 +38,18 @@ func deleteFilesAndEmptyDirs(path string, logFile *os.File) error {
 		}
 	}
 
-	if err := os.Remove(path); err != nil {
-		log.Printf("Failed to delete directory: %s, error: %s", path, err)
-		return err
+	if path != rootPath {
+		if err := os.Remove(path); err != nil {
+			log.Printf("Failed to delete directory: %s, error: %s", path, err)
+			return err
+		}
 	}
-	log.Printf("Deleted directory %s", path)
 
 	return nil
 }
 
 func Run() {
+
 	logFile, err := os.OpenFile("/tmp/mediorum/reaper.txt", os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0644)
 	if err != nil {
 		fmt.Printf("Error opening log file: %v\n", err)
@@ -56,5 +57,14 @@ func Run() {
 	}
 	defer logFile.Close()
 
-	deleteFilesAndEmptyDirs("/file_storage", logFile)
+	log.SetOutput(logFile)
+
+	if _, exists := os.LookupEnv("NO_LEGACY_REAPER"); exists || strings.HasSuffix(os.Getenv("creatorNodeEndpoint"), ".audius.co") {
+		log.Printf("Skipping deletion due to NO_LEGACY_REAPER env var present.")
+		log.Printf("Sleeping.")
+		time.Sleep(10000 * time.Hour)
+		return
+	}
+
+	deleteFilesAndEmptyDirs(rootPath, logFile)
 }
